chore: update installer instructions in README
Prior to this commit we had temporarily linked to a specific
version of the installers in the README, as we were rolling out
some changes to the release process that caused intermediate
breakage in the latest releases.

Now that the release process has stabilized, this commit updates
the docs back to guide users to the latest release.  We also
include more information about the new installer formats available
for linux and windows, and reduce some repetition in the README.