POC: docusaurus plugin that supports injecting example snippets
This commit shows the beginnings of how we could inject example
snippets from the SDK github repos into the docs.  For now it
just uses a hard-coded map of snippet content but this can easily
be extended to pull in the content from the appropriate files
in the various repos.