chore: remove snippet id enum to make it easier to add snippets
Prior to this commit, there was enum that contained an authoritative
list of all known snippets.  Thus, to use an instance of one of our
react components that pulls in the snippet code from the SDK repos,
you needed to register the new snippet in the enum first.

My original thinking behind this was that it would make it easier
to do an audit of which languages were missing support for which
snippets, but now I believe that I can achieve that without requiring
this enum.  And removing the enum should make it quite a bit simpler
for devs to inject new snippets from the various SDK repos over time.