feat: add remaining sdk example tags to API reference (#245)
This commit adds <SdkExampleTabs> tags for all of the APIs that
didn't already have them. None of the SDKs actually have examples
for these yet, so these will not actually render on the website
for now. But this allows us to start adding them to the example
code in the various SDKs.