fix: remove generics from function signatures when building api matrix (#265)
This commit removes generic type info from function signatures before
we try to populate the API support matrix. In some languages (at least
Rust) there is a generic type clause (`<...>`) after the function
name but before the opening parens and this was causing the generator
to miss some APIs in Rust that are actually implemented.