chore: add examples for new JS config options, other minor cleanups
This commit adds a "config-and-error-handling" page for the node.js
SDK, and in there we show some examples of how to use the succinct
version of the constructor without specifying an explicit config.
We also show an example of using the new `WithThrowOnErrors` configuration
option.

Also did a few other cleanups while I was in here, including:

* add missing example code for Lambda configuration on Configurations page.
* Remove all the "COMING SOON"s
* Fixed the order of the `TopicClient` example on the topics API reference page.
* Removed some redundant copies of stuff like 'observability' that was duplicated
  between "Topics" and "Cache" sections; but then gave up on this in favor of
  moving the "SDKs" stuff up to the top level in a subsequent PR after discussion
  with Allen.