chore: move cache `Integrations` section up out of `develop` (#1034)
chore: move cache 'Integrations' section up out of 'Develop'

On the Topics page, we have `Integrations` as a top level nav
item. But for some reason on the Cache page it was nested inside
`Develop`.

This commit moves /cache/develop/integrations up to just /cache/integrations
and updates the side nav accordingly, to make it consistent
with Topics.